Merge

+++ lib/util-disk.sh
@@ -223,6 +223,7 @@ confirm_mount() {
     else
         DIALOG " $_MntStatusTitle " --infobox "\n$_MntStatusFail\n " 0 0
         sleep 2
+        return 1
     fi
 }
 
@@ -298,7 +299,8 @@ select_filesystem() {
         DIALOG " $_FSTitle " --yesno "\n$_FSMount $FILESYSTEM\n\n! $_FSWarn1 $PARTITION $_FSWarn2 !\n " 0 0
         if (( $? != 1 )); then
             ${FILESYSTEM} ${PARTITION} >/dev/null 2>$ERR
-            check_for_error "mount $PARTITION as $FILESYSTEM." $?
+            check_for_error "mount ${PARTITION} as ${FILESYSTEM}." $? || return 1
+            ini "mount.${PARTITION}" $(echo "${FILESYSTEM:5}|cut -d' ' -f1")
         fi
     fi
 }
@@ -449,22 +451,31 @@ make_swap() {
             NUMBER_PARTITIONS=$(( NUMBER_PARTITIONS - 1 ))
         fi
     fi
+    ini mount.swap "${PARTITION}"
 }
 
-# Had to write it in this way due to (bash?) bug(?), as if/then statements in a single
-# "create LUKS" function for default and "advanced" modes were interpreted as commands,
-# not mere string statements. Not happy with it, but it works...
-luks_password() {
-    DIALOG " $_PrepLUKS " --clear --insecure --passwordbox "\n$_LuksPassBody\n " 0 0 2> ${ANSWER} || return 0
-    PASSWD=$(cat ${ANSWER})
+luks_menu() {
+    declare -i loopmenu=1
+    while ((loopmenu)); do
+        LUKS_OPT=""
+        DIALOG " $_PrepLUKS " --menu "\n$_LuksMenuBody\n$_LuksMenuBody2\n$_LuksMenuBody3\n " 0 0 4 \
+          "$_LuksOpen" "cryptsetup open --type luks" \
+          "$_LuksEncrypt" "cryptsetup -q luksFormat" \
+          "$_LuksEncryptAdv" "cryptsetup -q -s -c luksFormat" \
+          "$_Back" "-" 2>${ANSWER}
 
-    DIALOG " $_PrepLUKS " --clear --insecure --passwordbox "\n$_PassReEntBody\n " 0 0 2> ${ANSWER} || return 0
-    PASSWD2=$(cat ${ANSWER})
-
-    if [[ $PASSWD != $PASSWD2 ]]; then
-        DIALOG " $_ErrTitle " --msgbox "\n$_PassErrBody\n " 0 0
-        luks_password
-    fi
+        case $(cat ${ANSWER}) in
+            "$_LuksOpen") luks_open
+                ;;
+            "$_LuksEncrypt") luks_setup && luks_default && luks_show
+                ;;
+            "$_LuksEncryptAdv") luks_setup && luks_key_define && luks_show
+                ;;
+            *) loopmenu=0
+               return 0
+                ;;
+        esac
+    done
 }
 
 luks_open() {
@@ -479,14 +490,21 @@ luks_open() {
         delete_partition_in_list $part
     done
 
+    # stop if no encrypted partition found
+    if [[ $PARTITIONS == "" ]]; then
+        DIALOG " $_ErrTitle " --msgbox "\n$_LuksErr\n " 0 0
+        return 1
+    fi
+
     # Select encrypted partition to open
-    DIALOG " $_LuksOpen " --menu "\n$_LuksMenuBody\n " 0 0 12 ${PARTITIONS} 2>${ANSWER} || return 1
+    DIALOG " $_LuksOpen " --menu "\n$_LuksMenuBody\n " 0 0 12 ${PARTITIONS} 2>${ANSWER} || return 0
     PARTITION=$(cat ${ANSWER})
 
     # Enter name of the Luks partition and get password to open it
-    DIALOG " $_LuksOpen " --inputbox "\n$_LuksOpenBody\n " 10 50 "cryptroot" 2>${ANSWER} || return 1
+    DIALOG " $_LuksOpen " --inputbox "\n$_LuksOpenBody\n " 0 0 "cryptroot" 2>${ANSWER} || return 0
     LUKS_ROOT_NAME=$(cat ${ANSWER})
-    luks_password
+    DIALOG " $_PrepLUKS " --clear --insecure --passwordbox "\n$_LuksPassBody\n " 0 0 2> ${ANSWER} || return 0
+    PASSWD=$(cat ${ANSWER})
 
     # Try to open the luks partition with the credentials given. If successful show this, otherwise
     # show the error
@@ -494,8 +512,24 @@ luks_open() {
     echo $PASSWD | cryptsetup open --type luks ${PARTITION} ${LUKS_ROOT_NAME} 2>$ERR
     check_for_error "luks pwd ${PARTITION} ${LUKS_ROOT_NAME}" "$?"
 
-    lsblk -o NAME,TYPE,FSTYPE,SIZE,MOUNTPOINT ${PARTITION} | grep "crypt\|NAME\|MODEL\|TYPE\|FSTYPE\|SIZE" > /tmp/.devlist
+    echo "" > /tmp/.devlist
+    lsblk -o NAME,TYPE,FSTYPE,SIZE,MOUNTPOINT ${PARTITION} | grep "crypt\|NAME\|MODEL\|TYPE\|FSTYPE\|SIZE" >> /tmp/.devlist
     DIALOG " $_DevShowOpt " --textbox /tmp/.devlist 0 0
+}
+
+luks_password() {
+    luks_get_password
+    while [[ $PASSWD != $PASSWD2 ]]; do
+        DIALOG " $_ErrTitle " --msgbox "\n$_PassErrBody\n " 0 0
+        luks_get_password
+    done
+}
+
+luks_get_password() {
+    DIALOG " $_PrepLUKS " --clear --insecure --passwordbox "\n$_LuksPassBody\n " 0 0 2> ${ANSWER} || return 0
+    PASSWD=$(cat ${ANSWER})
+    DIALOG " $_PrepLUKS " --clear --insecure --passwordbox "\n$_PassReEntBody\n " 0 0 2> ${ANSWER} || return 0
+    PASSWD2=$(cat ${ANSWER})
 }
 
 luks_setup() {
@@ -508,7 +542,7 @@ luks_setup() {
     PARTITION=$(cat ${ANSWER})
 
     # Enter name of the Luks partition and get password to create it
-    DIALOG " $_LuksEncrypt " --inputbox "\n$_LuksOpenBody\n " 10 50 "cryptroot" 2>${ANSWER} || return 1
+    DIALOG " $_LuksEncrypt " --inputbox "\n$_LuksOpenBody\n " 0 0 "cryptroot" 2>${ANSWER} || return 1
     LUKS_ROOT_NAME=$(cat ${ANSWER})
     luks_password
 }
@@ -541,30 +575,36 @@ luks_key_define() {
 }
 
 luks_show() {
-    echo -e ${_LuksEncruptSucc} > /tmp/.devlist
+    printf "\n${_LuksEncruptSucc}\n\n" > /tmp/.devlist
     lsblk -o NAME,TYPE,FSTYPE,SIZE ${PARTITION} | grep "part\|crypt\|NAME\|TYPE\|FSTYPE\|SIZE" >> /tmp/.devlist
     DIALOG " $_LuksEncrypt " --textbox /tmp/.devlist 0 0
 }
 
-luks_menu() {
-    LUKS_OPT=""
+lvm_menu() {
+    declare -i loopmenu=1
+    while ((loopmenu)); do
+        DIALOG " $_PrepLVM $_PrepLVM2 " --infobox "\n$_PlsWaitBody\n " 0 0
+        sleep 1
+        lvm_detect
 
-    DIALOG " $_PrepLUKS " --menu "\n$_LuksMenuBody$_LuksMenuBody2$_LuksMenuBody3\n " 0 0 4 \
-      "$_LuksOpen" "cryptsetup open --type luks" \
-      "$_LuksEncrypt" "cryptsetup -q luksFormat" \
-      "$_LuksEncryptAdv" "cryptsetup -q -s -c luksFormat" \
-      "$_Back" "-" 2>${ANSWER}
+        DIALOG " $_PrepLVM $_PrepLVM2 " --menu "\n$_LvmMenu\n " 0 0 4 \
+          "$_LvmCreateVG" "vgcreate -f, lvcreate -L -n" \
+          "$_LvmDelVG" "vgremove -f" \
+          "$_LvMDelAll" "lvrmeove, vgremove, pvremove -f" \
+          "$_Back" "-" 2>${ANSWER}
 
-    case $(cat ${ANSWER}) in
-        "$_LuksOpen") luks_open
-            ;;
-        "$_LuksEncrypt") luks_setup && luks_default && luks_show
-            ;;
-            "$_LuksEncryptAdv") luks_setup && luks_key_define && luks_show
-            ;;
-        *) return 0
-            ;;
-    esac
+        case $(cat ${ANSWER}) in
+            "$_LvmCreateVG") lvm_create
+               ;;
+            "$_LvmDelVG") lvm_del_vg
+               ;;
+            "$_LvMDelAll") lvm_del_all
+               ;;
+            *) loopmenu=0
+               return 0
+               ;;
+        esac
+    done
 }
 
 lvm_detect() {
@@ -574,20 +614,12 @@ lvm_detect() {
 
     if [[ $LVM_LV != "" ]] && [[ $LVM_VG != "" ]] && [[ $LVM_PV != "" ]]; then
         DIALOG " $_PrepLVM " --infobox "\n$_LvmDetBody\n " 0 0
+        sleep 2
         modprobe dm-mod 2>$ERR
         check_for_error "modprobe dm-mod" "$?"
         vgscan >/dev/null 2>&1
         vgchange -ay >/dev/null 2>&1
     fi
-}
-
-lvm_show_vg() {
-    VG_LIST=""
-    vg_list=$(lvs --noheadings | awk '{print $2}' | uniq)
-
-    for i in ${vg_list}; do
-        VG_LIST="${VG_LIST} ${i} $(vgdisplay ${i} | grep -i "vg size" | awk '{print $3$4}')"
-    done
 }
 
 # Create Volume Group and Logical Volumes
@@ -600,7 +632,7 @@ lvm_create() {
     PARTITIONS=$(echo $PARTITIONS | sed 's/M\|G\|T/& off/g')
 
     # Name the Volume Group
-    DIALOG " $_LvmCreateVG " --inputbox "\n$_LvmNameVgBody\n " 0 0 "" 2>${ANSWER} || return 0
+    DIALOG " $_LvmCreateVG " --inputbox "\n$_LvmNameVgBody\n " 0 0 2>${ANSWER} || return 0
     LVM_VG=$(cat ${ANSWER})
 
     # Loop while the Volume Group name starts with a "/", is blank, has spaces, or is already being used
@@ -611,27 +643,31 @@ lvm_create() {
     done
 
     # Select the partition(s) for the Volume Group
-    DIALOG " $_LvmCreateVG " --checklist "\n$_LvmPvSelBody\n\n$_UseSpaceBar\n " 0 0 12 ${PARTITIONS} 2>${ANSWER} || return 0
-    [[ $(cat ${ANSWER}) != "" ]] && VG_PARTS=$(cat ${ANSWER}) || return 0
+    echo "" > $ANSWER
+    while [[ $(cat ${ANSWER}) == "" ]]; do
+        DIALOG " $_LvmCreateVG " --checklist "\n$_LvmPvSelBody\n\n$_UseSpaceBar\n " 0 0 12 ${PARTITIONS} 2>${ANSWER} || return 0
+    done
+
+    VG_PARTS=$(cat ${ANSWER})
 
     # Once all the partitions have been selected, show user. On confirmation, use it/them in 'vgcreate' command.
     # Also determine the size of the VG, to use for creating LVs for it.
-    DIALOG " $_LvmCreateVG " --yesno "\n$_LvmPvConfBody1${LVM_VG} $_LvmPvConfBody2${VG_PARTS}\n " 0 0
+    DIALOG " $_LvmCreateVG " --yesno "\n$_LvmPvConfBody1 [${LVM_VG}] $_LvmPvConfBody2\n${VG_PARTS}\n " 0 0
 
     if [[ $? -eq 0 ]]; then
-        DIALOG " $_LvmCreateVG " --infobox "\n$_LvmPvActBody1${LVM_VG}.$_PlsWaitBody\n " 0 0
+        DIALOG " $_LvmCreateVG " --infobox "\n$_LvmPvActBody1 [${LVM_VG}].\n$_PlsWaitBody\n " 0 0
         sleep 1
-        vgcreate -f ${LVM_VG} ${VG_PARTS} >/dev/null 2>$ERR
-        check_for_error "vgcreate -f ${LVM_VG} ${VG_PARTS}" "$?"
+        vgcreate -f ${LVM_VG} ${VG_PARTS} >/dev/null
+        check_for_error "vgcreate -f ${LVM_VG} ${VG_PARTS}"
 
         # Once created, get size and size type for display and later number-crunching for lv creation
-        VG_SIZE=$(vgdisplay $LVM_VG | grep 'VG Size' | awk '{print $3}' | sed 's/\..*//')
+        VG_SIZE=$(vgdisplay $LVM_VG | grep 'VG Size' | awk '{print $3}' | sed 's/\..*//; s/\,.*//')
         VG_SIZE_TYPE=$(vgdisplay $LVM_VG | grep 'VG Size' | awk '{print $4}')
 
         # Convert the VG size into GB and MB. These variables are used to keep tabs on space available and remaining
         [[ ${VG_SIZE_TYPE:0:1} == "G" ]] && LVM_VG_MB=$(( VG_SIZE * 1000 )) || LVM_VG_MB=$VG_SIZE
 
-        DIALOG " $_LvmCreateVG " --msgbox "\n$_LvmPvDoneBody1 '${LVM_VG}' $_LvmPvDoneBody2 (${VG_SIZE} ${VG_SIZE_TYPE}).\n " 0 0 || return 0
+        DIALOG " $_LvmCreateVG " --msgbox "\n$_LvmPvDoneBody1 [${LVM_VG}] (${VG_SIZE} ${VG_SIZE_TYPE}) $_LvmPvDoneBody2.\n " 0 0 || return 0
     fi
 
     #
@@ -639,10 +675,14 @@ lvm_create() {
     #
 
     # Specify number of Logical volumes to create.
-    DIALOG " $_LvmCreateVG " --radiolist "\n$_LvmLvNumBody1 ${LVM_VG}. $_LvmLvNumBody2\n " 0 0 9 \
-      "1" "-" off "2" "-" off "3" "-" off "4" "-" off "5" "-" off "6" "-" off "7" "-" off "8" "-" off "9" "-" off 2>${ANSWER}
+    DIALOG " $_LvmCreateVG " --inputbox "\n$_LvmLvNumBody1 [${LVM_VG}]. $_LvmLvNumBody2\n " 0 0 2>${ANSWER}
 
-    [[ $(cat ${ANSWER}) == "" ]] && lvm_menu || NUMBER_LOGICAL_VOLUMES=$(cat ${ANSWER})
+    # repeat if answer is not a number
+    while [[ $(cat ${ANSWER}) != ?(-)+([0-9]) ]]; do
+        DIALOG " $_ErrTitle " --inputbox "\n$_LvmLvNumBody1 [${LVM_VG}]. $_LvmLvNumBody2\n " 0 0 2>${ANSWER}
+    done
+
+    NUMBER_LOGICAL_VOLUMES=$(cat ${ANSWER})
 
     # Loop while the number of LVs is greater than 1. This is because the size of the last LV is automatic.
     while [[ $NUMBER_LOGICAL_VOLUMES -gt 1 ]]; do
@@ -656,16 +696,16 @@ lvm_create() {
             LVM_LV_NAME=$(cat ${ANSWER})
         done
 
-        DIALOG " $_LvmCreateVG (LV:$NUMBER_LOGICAL_VOLUMES) " --inputbox "\n${LVM_VG}: ${VG_SIZE}${VG_SIZE_TYPE} (${LVM_VG_MB}MB \
-          $_LvmLvSizeBody1).$_LvmLvSizeBody2\n " 0 0 "" 2>${ANSWER} || return 0
+        DIALOG " $_LvmCreateVG (LV:$NUMBER_LOGICAL_VOLUMES) " --inputbox \
+          "\n[${LVM_VG}]: ${VG_SIZE}${VG_SIZE_TYPE} - (${LVM_VG_MB}MB $_LvmLvSizeBody1).\n\n$_LvmLvSizeBody2\n " 0 0 "" 2>${ANSWER} || return 0
         LVM_LV_SIZE=$(cat ${ANSWER})
         check_lv_size
 
         # Loop while an invalid value is entered.
         while [[ $LV_SIZE_INVALID -eq 1 ]]; do
             DIALOG " $_ErrTitle " --msgbox "\n$_LvmLvSizeErrBody\n " 0 0
-            DIALOG " $_LvmCreateVG (LV:$NUMBER_LOGICAL_VOLUMES) " --inputbox "\n${LVM_VG}: ${VG_SIZE}${VG_SIZE_TYPE} \
-              (${LVM_VG_MB}MB $_LvmLvSizeBody1).$_LvmLvSizeBody2\n " 0 0 "" 2>${ANSWER} || return 0
+            DIALOG " $_LvmCreateVG (LV:$NUMBER_LOGICAL_VOLUMES) " --inputbox \
+              "\n[${LVM_VG}]: ${VG_SIZE}${VG_SIZE_TYPE} - (${LVM_VG_MB}MB $_LvmLvSizeBody1).\n\n$_LvmLvSizeBody2\n " 0 0 "" 2>${ANSWER} || return 0
             LVM_LV_SIZE=$(cat ${ANSWER})
             check_lv_size
         done
@@ -673,7 +713,7 @@ lvm_create() {
         # Create the LV
         lvcreate -L ${LVM_LV_SIZE} ${LVM_VG} -n ${LVM_LV_NAME} 2>$ERR
         check_for_error "lvcreate -L ${LVM_LV_SIZE} ${LVM_VG} -n ${LVM_LV_NAME}" "$?"
-        DIALOG " $_LvmCreateVG (LV:$NUMBER_LOGICAL_VOLUMES) " --msgbox "\n$_Done\n\nLV ${LVM_LV_NAME} (${LVM_LV_SIZE}) $_LvmPvDoneBody2.\n " 0 0
+        DIALOG " $_LvmCreateVG (LV:$NUMBER_LOGICAL_VOLUMES) " --msgbox "\nLV ${LVM_LV_NAME} (${LVM_LV_SIZE}) $_LvmPvDoneBody2.\n " 0 0
         NUMBER_LOGICAL_VOLUMES=$(( NUMBER_LOGICAL_VOLUMES - 1 ))
     done
 
@@ -765,12 +805,27 @@ lvm_del_vg() {
 
     # if confirmation given, delete
     if [[ $? -eq 0 ]]; then
+        vgremove -f $(cat ${ANSWER}) 2>/dev/null
         check_for_error "delete lvm-VG $(cat ${ANSWER})"
-        vgremove -f $(cat ${ANSWER}) >/dev/null 2>&1
     fi
 }
 
+lvm_show_vg() {
+    VG_LIST=""
+    vg_list=$(lvs --noheadings | awk '{print $2}' | uniq)
+
+    for i in ${vg_list}; do
+        VG_LIST="${VG_LIST} ${i} $(vgdisplay ${i} | grep -i "vg size" | awk '{print $3$4}')"
+    done
+}
+
 lvm_del_all() {
+    # check if VG exist at all
+    if [[ $(lvs) == "" ]]; then
+        DIALOG " $_ErrTitle " --msgbox "\n$_LvmVGErr\n " 0 0
+        return 0
+    fi
+
     LVM_PV=$(pvs -o pv_name --noheading 2>/dev/null)
     LVM_VG=$(vgs -o vg_name --noheading 2>/dev/null)
     LVM_LV=$(lvs -o vg_name,lv_name --noheading --separator - 2>/dev/null)
@@ -781,39 +836,20 @@ lvm_del_all() {
     # if confirmation given, delete
     if [[ $? -eq 0 ]]; then
         for i in ${LVM_LV}; do
+            lvremove -f /dev/mapper/${i} 2>/dev/null
             check_for_error "remove LV ${i}"
-            lvremove -f /dev/mapper/${i} >/dev/null 2>&1
         done
 
         for i in ${LVM_VG}; do
+            vgremove -f ${i} 2>/dev/null
             check_for_error "remove VG ${i}"
-            vgremove -f ${i} >/dev/null 2>&1
         done
 
         for i in ${LV_PV}; do
+            pvremove -f ${i} 2>/dev/null
             check_for_error "remove LV-PV ${i}"
-            pvremove -f ${i} >/dev/null 2>&1
         done
     fi
-}
-
-lvm_menu() {
-    DIALOG " $_PrepLVM $_PrepLVM2 " --infobox "\n$_PlsWaitBody\n " 0 0
-    sleep 1
-    lvm_detect
-
-    DIALOG " $_PrepLVM $_PrepLVM2 " --menu "\n$_LvmMenu\n " 0 0 4 \
-      "$_LvmCreateVG" "vgcreate -f, lvcreate -L -n" \
-      "$_LvmDelVG" "vgremove -f" \
-      "$_LvMDelAll" "lvrmeove, vgremove, pvremove -f" \
-      "$_Back" "-" 2>${ANSWER}
-
-    case $(cat ${ANSWER}) in
-        "$_LvmCreateVG") lvm_create ;;
-        "$_LvmDelVG") lvm_del_vg ;;
-        "$_LvMDelAll") lvm_del_all ;;
-        *) return 0 ;;
-    esac
 }
 
 mount_partitions() {
@@ -844,37 +880,45 @@ mount_partitions() {
     # Format with FS (or skip) -> # Make the directory and mount. Also identify LUKS and/or LVM
     select_filesystem && mount_current_partition || return 0
 
+    ini mount.root "${PARTITION}"
+    delete_partition_in_list "${ROOT_PART}"
+
     # Identify and create swap, if applicable
     make_swap
 
     # Extra Step for VFAT UEFI Partition. This cannot be in an LVM container.
     if [[ $SYSTEM == "UEFI" ]]; then
-        DIALOG " $_PrepMntPart " --menu "\n$_SelUefiBody\n " 0 0 12 ${PARTITIONS} 2>${ANSWER} || return 0
-        PARTITION=$(cat ${ANSWER})
-        UEFI_PART=${PARTITION}
+        if DIALOG " $_PrepMntPart " --menu "\n$_SelUefiBody\n " 0 0 12 ${PARTITIONS} 2>${ANSWER}; then
+            PARTITION=$(cat ${ANSWER})
+            UEFI_PART=${PARTITION}
 
-        # If it is already a fat/vfat partition...
-        if [[ $(fsck -N $PARTITION | grep fat) ]]; then
-            DIALOG " $_PrepMntPart " --yesno "\n$_FormUefiBody $PARTITION $_FormUefiBody2\n " 0 0 && {
+            # If it is already a fat/vfat partition...
+            if [[ $(fsck -N $PARTITION | grep fat) ]]; then
+                DIALOG " $_PrepMntPart " --yesno "\n$_FormUefiBody $PARTITION $_FormUefiBody2\n " 0 0 && {
+                    mkfs.vfat -F32 ${PARTITION} >/dev/null 2>$ERR
+                    check_for_error "mkfs.vfat -F32 ${PARTITION}" "$?"
+                } # || return 0
+            else
                 mkfs.vfat -F32 ${PARTITION} >/dev/null 2>$ERR
                 check_for_error "mkfs.vfat -F32 ${PARTITION}" "$?"
-            }
-        else
-            mkfs.vfat -F32 ${PARTITION} >/dev/null 2>$ERR
-            check_for_error "mkfs.vfat -F32 ${PARTITION}" "$?"
+            fi
+
+            DIALOG " $_PrepMntPart " --radiolist "\n$_MntUefiBody\n "  0 0 2 \
+            "/boot" "" on \
+            "/boot/efi" "" off 2>${ANSWER}
+
+            if [[ $(cat ${ANSWER}) != "" ]]; then
+                UEFI_MOUNT=$(cat ${ANSWER})
+                mkdir -p ${MOUNTPOINT}${UEFI_MOUNT} 2>$ERR
+                check_for_error "create ${MOUNTPOINT}${UEFI_MOUNT}" $?
+                mount ${PARTITION} ${MOUNTPOINT}${UEFI_MOUNT} 2>$ERR
+                check_for_error "mount ${PARTITION} ${MOUNTPOINT}${UEFI_MOUNT}" $?
+                if confirm_mount ${MOUNTPOINT}${UEFI_MOUNT}; then
+                    ini mount.efi "${UEFI_MOUNT}"
+                    delete_partition_in_list "$PARTITION"
+                fi
+            fi
         fi
-
-        DIALOG " $_PrepMntPart " --radiolist "\n$_MntUefiBody\n "  0 0 2 \
-          "/boot" "" on \
-          "/boot/efi" "" off 2>${ANSWER}
-
-        [[ $(cat ${ANSWER}) != "" ]] && UEFI_MOUNT=$(cat ${ANSWER}) || return 0
-
-        mkdir -p ${MOUNTPOINT}${UEFI_MOUNT} 2>$ERR
-        check_for_error "create ${MOUNTPOINT}${UEFI_MOUNT}" "$?"
-        mount ${PARTITION} ${MOUNTPOINT}${UEFI_MOUNT} 2>$ERR
-        check_for_error "mount ${PARTITION} ${MOUNTPOINT}${UEFI_MOUNT}" "$?"
-        confirm_mount ${MOUNTPOINT}${UEFI_MOUNT}
     fi
 
     # All other partitions
@@ -904,6 +948,7 @@ mount_partitions() {
 
             # Create directory and mount.
             mount_current_partition
+            delete_partition_in_list "$PARTITION"
 
             # Determine if a seperate /boot is used. 0 = no seperate boot, 1 = seperate non-lvm boot,
             # 2 = seperate lvm boot. For Grub configuration
